requirements.txt
transformers==4.49.0
torch==2.6.0
pillow==11.1.0
googletrans==4.0.2
translate==3.6.1
SQLAlchemy==2.0.38
gradio==5.20.1
psycopg2-binary==2.9.10
click==8.1.8
python-dotenv==1.0.1
pyyaml==6.0.2
fastapi[standart]==0.115.11